+++ src/checks.c
@@ -7,17 +7,17 @@
 #include "all_missing.h"
 #include "all_nchar.h"
 #include "helper.h"
-#include "guessType.h"
+#include "guess_type.h"
 
 static char msg[255] = "";
 
-#define handle_type(expr, expected) \
+#define HANDLE_TYPE(expr, expected) \
     if (!(expr)) { \
-        snprintf(msg, 255, "Must be of type '%s', not '%s'", expected, guessType(x)); \
+        snprintf(msg, 255, "Must be of type '%s', not '%s'", expected, guess_type(x)); \
         return ScalarString(mkChar(msg)); \
     }
 
-#define handle_type_null(expr, expected, null_ok) \
+#define HANDLE_TYPE_NULL(expr, expected, null_ok) \
     if (isNull((x))) { \
         if (asFlag((null_ok), "null.ok")) \
             return ScalarLogical(TRUE); \
@@ -25,20 +25,19 @@ static char msg[255] = "";
         return ScalarString(mkChar(msg)); \
     } else { \
         if (!(expr)) { \
-            snprintf(msg, 255, "Must be of type '%s'%s, not '%s'", expected, asFlag(null_ok, "null_ok") ? " (or 'NULL')" : "", guessType(x)); \
+            snprintf(msg, 255, "Must be of type '%s'%s, not '%s'", expected, asFlag(null_ok, "null_ok") ? " (or 'NULL')" : "", guess_type(x)); \
             return ScalarString(mkChar(msg)); \
         } \
     }
 
-#define handle_na(x, na_ok) \
+#define HANDLE_NA(x, na_ok) \
     if (is_scalar_na((x))) { \
         if (asFlag((na_ok), "na.ok")) \
             return ScalarLogical(TRUE); \
         return result("May not be NA"); \
     };
 
-
-#define assert(x) if (!(x)) return ScalarString(mkChar(msg));
+#define ASSERT_TRUE(x) if (!(x)) return ScalarString(mkChar(msg));
 
 
 /*********************************************************************************************************************/
@@ -117,23 +116,21 @@ static Rboolean check_strict_names(SEXP x) {
     }
     return TRUE;
 }
-
-static Rboolean check_names(SEXP nn, SEXP type, const char * what) {
+static Rboolean check_names(SEXP nn, const char * type, const char * what) {
     typedef enum { T_NAMED, T_UNIQUE, T_STRICT } name_t;
-    const char * expected = asString(type, "names");
+    name_t checks;
 
-    if (strcmp(expected, "unnamed") == 0)
+    if (strcmp(type, "unnamed") == 0)
         return isNull(nn) ? TRUE : message("%s must be unnamed, but has names", what);
 
-    name_t checks;
-    if (strcmp(expected, "named") == 0) {
+    if (strcmp(type, "named") == 0) {
         checks = T_NAMED;
-    } else if (strcmp(expected, "unique") == 0) {
+    } else if (strcmp(type, "unique") == 0) {
         checks = T_UNIQUE;
-    } else if (strcmp(expected, "strict") == 0) {
+    } else if (strcmp(type, "strict") == 0) {
         checks = T_STRICT;
     } else {
-        error("Unknown type '%s' to specify check for names. Supported are 'unnamed', 'named', 'unique' and 'strict'.", expected);
+        error("Unknown type '%s' to specify check for names. Supported are 'unnamed', 'named', 'unique' and 'strict'.", type);
     }
 
     if (isNull(nn) || any_missing_string(nn) || !all_nchar(nn, 1))
@@ -182,7 +179,7 @@ static Rboolean check_vector_unique(SEXP x, SEXP unique) {
 
 static Rboolean check_vector_names(SEXP x, SEXP names) {
     if (!isNull(names) && xlength(x) > 0)
-        return check_names(getAttrib(x, R_NamesSymbol), names, "Vector");
+        return check_names(getAttrib(x, R_NamesSymbol), asString(names, "names"), "Vector");
     return TRUE;
 }
 
@@ -234,6 +231,9 @@ static Rboolean check_storage(SEXP x, SEXP mode) {
         } else if (strcmp(storage, "double") == 0) {
             if (!isReal(x))
                 return message("Must store doubles");
+        } else if (strcmp(storage, "integerish") == 0) {
+            if (!isIntegerish(x, INTEGERISH_DEFAULT_TOL, FALSE))
+                return message("Must store integerish values");
         } else if (strcmp(storage, "numeric") == 0) {
             if (!isStrictlyNumeric(x))
                 return message("Must store numerics");
@@ -250,7 +250,7 @@ static Rboolean check_storage(SEXP x, SEXP mode) {
             if (!isVectorAtomic(x))
                 return message("Must be atomic");
         } else {
-            error("Invalid argument 'mode'. Must be one of 'logical', 'integer', 'double', 'numeric', 'complex', 'character', 'list' or 'atomic'");
+            error("Invalid argument 'mode'. Must be one of 'logical', 'integer', 'integerish', 'double', 'numeric', 'complex', 'character', 'list' or 'atomic'");
         }
     }
     return TRUE;
@@ -273,47 +273,47 @@ static inline Rboolean is_scalar_na(SEXP x) {
 /* Exported check functions                                                                                          */
 /*********************************************************************************************************************/
 SEXP c_check_character(SEXP x, SEXP min_chars, SEXP any_missing, SEXP all_missing, SEXP len, SEXP min_len, SEXP max_len, SEXP unique, SEXP names, SEXP null_ok) {
-    handle_type_null(isString(x) || all_missing_atomic(x), "character", null_ok);
-    assert(check_vector_len(x, len, min_len, max_len));
-    assert(check_vector_names(x, names));
-    assert(check_vector_missings(x, any_missing, all_missing));
+    HANDLE_TYPE_NULL(isString(x) || all_missing_atomic(x), "character", null_ok);
+    ASSERT_TRUE(check_vector_len(x, len, min_len, max_len));
+    ASSERT_TRUE(check_vector_names(x, names));
+    ASSERT_TRUE(check_vector_missings(x, any_missing, all_missing));
     if (!isNull(min_chars)) {
         R_xlen_t n = asCount(min_chars, "min.chars");
         if (n > 0 && !all_nchar(x, n))
             return result("All elements must have at least %i characters", n);
     }
-    assert(check_vector_unique(x, unique));
+    ASSERT_TRUE(check_vector_unique(x, unique));
     return ScalarLogical(TRUE);
 }
 
 SEXP c_check_complex(SEXP x, SEXP any_missing, SEXP all_missing, SEXP len, SEXP min_len, SEXP max_len, SEXP unique, SEXP names, SEXP null_ok) {
-    handle_type_null(isComplex(x) || all_missing_atomic(x), "complex", null_ok);
-    assert(check_vector_len(x, len, min_len, max_len));
-    assert(check_vector_names(x, names));
-    assert(check_vector_missings(x, any_missing, all_missing));
-    assert(check_vector_unique(x, unique));
+    HANDLE_TYPE_NULL(isComplex(x) || all_missing_atomic(x), "complex", null_ok);
+    ASSERT_TRUE(check_vector_len(x, len, min_len, max_len));
+    ASSERT_TRUE(check_vector_names(x, names));
+    ASSERT_TRUE(check_vector_missings(x, any_missing, all_missing));
+    ASSERT_TRUE(check_vector_unique(x, unique));
     return ScalarLogical(TRUE);
 }
 
 SEXP c_check_dataframe(SEXP x, SEXP any_missing, SEXP all_missing, SEXP min_rows, SEXP min_cols, SEXP rows, SEXP cols, SEXP row_names, SEXP col_names, SEXP null_ok) {
-    handle_type_null(isFrame(x), "data.frame", null_ok);
-    assert(check_matrix_dims(x, min_rows, min_cols, rows, cols));
+    HANDLE_TYPE_NULL(isFrame(x), "data.frame", null_ok);
+    ASSERT_TRUE(check_matrix_dims(x, min_rows, min_cols, rows, cols));
 
     if (!isNull(row_names)) {
         SEXP nn = getAttrib(x, install("row.names"));
         if (isInteger(nn)) {
             nn = PROTECT(coerceVector(nn, STRSXP));
-            Rboolean ok = check_names(nn, row_names, "Rows");
+            Rboolean ok = check_names(nn, asString(row_names, "row.names"), "Rows");
             UNPROTECT(1);
             if (!ok)
                 return ScalarString(mkChar(msg));
         } else {
-            assert(check_names(nn, row_names, "Rows"));
+            ASSERT_TRUE(check_names(nn, asString(row_names, "row.names"), "Rows"));
         }
     }
 
     if (!isNull(col_names))
-        assert(check_names(getAttrib(x, R_NamesSymbol), col_names, "Columns"));
+        ASSERT_TRUE(check_names(getAttrib(x, R_NamesSymbol), asString(col_names, "col.names"), "Columns"));
     if (!asFlag(any_missing, "any.missing") && any_missing_frame(x))
         return result("Contains missing values");
     if (!asFlag(all_missing, "all.missing") && all_missing_frame(x))
@@ -322,78 +322,78 @@ SEXP c_check_dataframe(SEXP x, SEXP any_missing, SEXP all_missing, SEXP min_rows
 }
 
 SEXP c_check_factor(SEXP x, SEXP any_missing, SEXP all_missing, SEXP len, SEXP min_len, SEXP max_len, SEXP unique, SEXP names, SEXP null_ok) {
-    handle_type_null(isFactor(x) || all_missing_atomic(x), "factor", null_ok);
-    assert(check_vector_len(x, len, min_len, max_len));
-    assert(check_vector_names(x, names));
-    assert(check_vector_missings(x, any_missing, all_missing));
-    assert(check_vector_unique(x, unique));
+    HANDLE_TYPE_NULL(isFactor(x) || all_missing_atomic(x), "factor", null_ok);
+    ASSERT_TRUE(check_vector_len(x, len, min_len, max_len));
+    ASSERT_TRUE(check_vector_names(x, names));
+    ASSERT_TRUE(check_vector_missings(x, any_missing, all_missing));
+    ASSERT_TRUE(check_vector_unique(x, unique));
     return ScalarLogical(TRUE);
 }
 
 SEXP c_check_integer(SEXP x, SEXP lower, SEXP upper, SEXP any_missing, SEXP all_missing, SEXP len, SEXP min_len, SEXP max_len, SEXP unique, SEXP names, SEXP null_ok) {
-    handle_type_null(isInteger(x) || all_missing_atomic(x), "integer", null_ok);
-    assert(check_vector_len(x, len, min_len, max_len));
-    assert(check_vector_names(x, names));
-    assert(check_vector_missings(x, any_missing, all_missing));
-    assert(check_bounds(x, lower, upper));
-    assert(check_vector_unique(x, unique));
+    HANDLE_TYPE_NULL(isInteger(x) || all_missing_atomic(x), "integer", null_ok);
+    ASSERT_TRUE(check_vector_len(x, len, min_len, max_len));
+    ASSERT_TRUE(check_vector_names(x, names));
+    ASSERT_TRUE(check_vector_missings(x, any_missing, all_missing));
+    ASSERT_TRUE(check_bounds(x, lower, upper));
+    ASSERT_TRUE(check_vector_unique(x, unique));
     return ScalarLogical(TRUE);
 }
 
 SEXP c_check_integerish(SEXP x, SEXP tol, SEXP lower, SEXP upper, SEXP any_missing, SEXP all_missing, SEXP len, SEXP min_len, SEXP max_len, SEXP unique, SEXP names, SEXP null_ok) {
     double dtol = asNumber(tol, "tol");
-    handle_type_null(isIntegerish(x, dtol) || all_missing_atomic(x), "integerish", null_ok);
-    assert(check_vector_len(x, len, min_len, max_len));
-    assert(check_vector_names(x, names));
-    assert(check_vector_missings(x, any_missing, all_missing));
-    assert(check_bounds(x, lower, upper));
-    assert(check_vector_unique(x, unique));
+    HANDLE_TYPE_NULL(isIntegerish(x, dtol, FALSE) || all_missing_atomic(x), "integerish", null_ok);
+    ASSERT_TRUE(check_vector_len(x, len, min_len, max_len));
+    ASSERT_TRUE(check_vector_names(x, names));
+    ASSERT_TRUE(check_vector_missings(x, any_missing, all_missing));
+    ASSERT_TRUE(check_bounds(x, lower, upper));
+    ASSERT_TRUE(check_vector_unique(x, unique));
     return ScalarLogical(TRUE);
 }
 
 SEXP c_check_list(SEXP x, SEXP any_missing, SEXP all_missing, SEXP len, SEXP min_len, SEXP max_len, SEXP unique, SEXP names, SEXP null_ok) {
-    handle_type_null(isRList(x), "list", null_ok)
-    assert(check_vector_len(x, len, min_len, max_len));
-    assert(check_vector_names(x, names));
-    assert(check_vector_missings(x, any_missing, all_missing));
-    assert(check_vector_unique(x, unique));
+    HANDLE_TYPE_NULL(isRList(x), "list", null_ok)
+    ASSERT_TRUE(check_vector_len(x, len, min_len, max_len));
+    ASSERT_TRUE(check_vector_names(x, names));
+    ASSERT_TRUE(check_vector_missings(x, any_missing, all_missing));
+    ASSERT_TRUE(check_vector_unique(x, unique));
     return ScalarLogical(TRUE);
 }
 
 SEXP c_check_logical(SEXP x, SEXP any_missing, SEXP all_missing, SEXP len, SEXP min_len, SEXP max_len, SEXP unique, SEXP names, SEXP null_ok) {
-    handle_type_null(isLogical(x) || all_missing_atomic(x), "logical", null_ok);
-    assert(check_vector_len(x, len, min_len, max_len));
-    assert(check_vector_names(x, names));
-    assert(check_vector_missings(x, any_missing, all_missing));
-    assert(check_vector_unique(x, unique));
+    HANDLE_TYPE_NULL(isLogical(x) || all_missing_atomic(x), "logical", null_ok);
+    ASSERT_TRUE(check_vector_len(x, len, min_len, max_len));
+    ASSERT_TRUE(check_vector_names(x, names));
+    ASSERT_TRUE(check_vector_missings(x, any_missing, all_missing));
+    ASSERT_TRUE(check_vector_unique(x, unique));
     return ScalarLogical(TRUE);
 }
 
 SEXP c_check_matrix(SEXP x, SEXP mode, SEXP any_missing, SEXP all_missing, SEXP min_rows, SEXP min_cols, SEXP rows, SEXP cols, SEXP row_names, SEXP col_names, SEXP null_ok) {
-    handle_type_null(isMatrix(x), "matrix", null_ok);
-    assert(check_storage(x, mode));
-    assert(check_matrix_dims(x, min_rows, min_cols, rows, cols));
+    HANDLE_TYPE_NULL(isMatrix(x), "matrix", null_ok);
+    ASSERT_TRUE(check_storage(x, mode));
+    ASSERT_TRUE(check_matrix_dims(x, min_rows, min_cols, rows, cols));
 
     if (!isNull(row_names) && xlength(x) > 0) {
         SEXP nn = getAttrib(x, R_DimNamesSymbol);
         if (!isNull(nn))
             nn = VECTOR_ELT(nn, 0);
-        assert(check_names(nn, row_names, "Rows"));
+        ASSERT_TRUE(check_names(nn, asString(row_names, "row.names"), "Rows"));
     }
 
     if (!isNull(col_names) && xlength(x) > 0) {
         SEXP nn = getAttrib(x, R_DimNamesSymbol);
         if (!isNull(nn))
             nn = VECTOR_ELT(nn, 1);
-        assert(check_names(nn, col_names, "Columns"));
+        ASSERT_TRUE(check_names(nn, asString(col_names, "col.names"), "Columns"));
     }
-    assert(check_vector_missings(x, any_missing, all_missing));
+    ASSERT_TRUE(check_vector_missings(x, any_missing, all_missing));
     return ScalarLogical(TRUE);
 }
 
 SEXP c_check_array(SEXP x, SEXP mode, SEXP any_missing, SEXP d, SEXP min_d, SEXP max_d, SEXP null_ok) {
-    handle_type_null(isArray(x), "array", null_ok);
-    assert(check_storage(x, mode));
+    HANDLE_TYPE_NULL(isArray(x), "array", null_ok);
+    ASSERT_TRUE(check_storage(x, mode));
 
     if (!asFlag(any_missing, "any.missing") && any_missing_atomic(x))
         return result("Contains missing values");
@@ -422,72 +422,72 @@ SEXP c_check_array(SEXP x, SEXP mode, SEXP any_missing, SEXP d, SEXP min_d, SEXP
 
 SEXP c_check_named(SEXP x, SEXP type) {
     if (!isNull(type) && xlength(x) > 0)
-        assert(check_names(getAttrib(x, R_NamesSymbol), type, "Object"));
+        ASSERT_TRUE(check_names(getAttrib(x, R_NamesSymbol), asString(type, "type"), "Object"));
     return ScalarLogical(TRUE);
 }
 
 SEXP c_check_names(SEXP x, SEXP type) {
     if (!isString(x))
         return result("Must be a character vector of names");
-    assert(check_names(x, type, "Object"));
+    ASSERT_TRUE(check_names(x, asString(type, "type"), "Names"));
     return ScalarLogical(TRUE);
 }
 
 
 SEXP c_check_numeric(SEXP x, SEXP lower, SEXP upper, SEXP finite, SEXP any_missing, SEXP all_missing, SEXP len, SEXP min_len, SEXP max_len, SEXP unique, SEXP names, SEXP null_ok) {
-    handle_type_null(isNumeric(x) || all_missing_atomic(x), "numeric", null_ok);
-    assert(check_vector_len(x, len, min_len, max_len));
-    assert(check_vector_names(x, names));
-    assert(check_vector_missings(x, any_missing, all_missing));
-    assert(check_bounds(x, lower, upper));
-    assert(check_vector_finite(x, finite));
-    assert(check_vector_unique(x, unique));
+    HANDLE_TYPE_NULL(isStrictlyNumeric(x) || all_missing_atomic(x), "numeric", null_ok);
+    ASSERT_TRUE(check_vector_len(x, len, min_len, max_len));
+    ASSERT_TRUE(check_vector_names(x, names));
+    ASSERT_TRUE(check_vector_missings(x, any_missing, all_missing));
+    ASSERT_TRUE(check_bounds(x, lower, upper));
+    ASSERT_TRUE(check_vector_finite(x, finite));
+    ASSERT_TRUE(check_vector_unique(x, unique));
     return ScalarLogical(TRUE);
 }
 
 SEXP c_check_vector(SEXP x, SEXP strict, SEXP any_missing, SEXP all_missing, SEXP len, SEXP min_len, SEXP max_len, SEXP unique, SEXP names, SEXP null_ok) {
-    handle_type_null(isVector(x), "vector", null_ok);
+    HANDLE_TYPE_NULL(isVector(x), "vector", null_ok);
     if (asFlag(strict, "strict")) {
         SEXP attr = ATTRIB(x);
-        handle_type( (length(attr) == 0 || (TAG(attr) == R_NamesSymbol)) && CDR(attr) == R_NilValue, "vector");
+        HANDLE_TYPE( (length(attr) == 0 || (TAG(attr) == R_NamesSymbol)) && CDR(attr) == R_NilValue, "vector");
     }
-    assert(check_vector_len(x, len, min_len, max_len));
-    assert(check_vector_names(x, names));
-    assert(check_vector_missings(x, any_missing, all_missing));
-    assert(check_vector_unique(x, unique));
+    ASSERT_TRUE(check_vector_len(x, len, min_len, max_len));
+    ASSERT_TRUE(check_vector_names(x, names));
+    ASSERT_TRUE(check_vector_missings(x, any_missing, all_missing));
+    ASSERT_TRUE(check_vector_unique(x, unique));
     return ScalarLogical(TRUE);
 }
 
 SEXP c_check_atomic(SEXP x, SEXP any_missing, SEXP all_missing, SEXP len, SEXP min_len, SEXP max_len, SEXP unique, SEXP names) {
-    handle_type(isNull(x) || isVectorAtomic(x), "atomic");
-    assert(check_vector_len(x, len, min_len, max_len));
-    assert(check_vector_names(x, names));
-    assert(check_vector_missings(x, any_missing, all_missing));
-    assert(check_vector_unique(x, unique));
+    HANDLE_TYPE(isNull(x) || isVectorAtomic(x), "atomic");
+    ASSERT_TRUE(check_vector_len(x, len, min_len, max_len));
+    ASSERT_TRUE(check_vector_names(x, names));
+    ASSERT_TRUE(check_vector_missings(x, any_missing, all_missing));
+    ASSERT_TRUE(check_vector_unique(x, unique));
     return ScalarLogical(TRUE);
 }
 
 SEXP c_check_atomic_vector(SEXP x, SEXP any_missing, SEXP all_missing, SEXP len, SEXP min_len, SEXP max_len, SEXP unique, SEXP names) {
-    handle_type(isVectorAtomic(x), "atomic vector");
-    assert(check_vector_len(x, len, min_len, max_len));
-    assert(check_vector_names(x, names));
-    assert(check_vector_missings(x, any_missing, all_missing));
-    assert(check_vector_unique(x, unique));
+    HANDLE_TYPE(isAtomicVector(x), "atomic vector");
+    ASSERT_TRUE(check_vector_len(x, len, min_len, max_len));
+    ASSERT_TRUE(check_vector_names(x, names));
+    ASSERT_TRUE(check_vector_missings(x, any_missing, all_missing));
+    ASSERT_TRUE(check_vector_unique(x, unique));
     return ScalarLogical(TRUE);
 }
 
 SEXP c_check_flag(SEXP x, SEXP na_ok, SEXP null_ok) {
-    handle_na(x, na_ok);
-    handle_type_null(isLogical(x), "logical flag", null_ok);
+    HANDLE_NA(x, na_ok);
+    HANDLE_TYPE_NULL(isLogical(x), "logical flag", null_ok);
     if (xlength(x) != 1)
         return result("Must have length 1");
     return ScalarLogical(TRUE);
 }
 
 SEXP c_check_count(SEXP x, SEXP na_ok, SEXP positive, SEXP tol, SEXP null_ok) {
-    handle_na(x, na_ok)
+    HANDLE_NA(x, na_ok)
     double dtol = asNumber(tol, "tol");
-    handle_type_null(isIntegerish(x, dtol), "count", null_ok);
+    HANDLE_TYPE_NULL(isIntegerish(x, dtol, FALSE), "count", null_ok);
     if (xlength(x) != 1)
         return result("Must have length 1");
     const int pos = (int) asFlag(positive, "positive");
@@ -498,27 +498,27 @@ SEXP c_check_count(SEXP x, SEXP na_ok, SEXP positive, SEXP tol, SEXP null_ok) {
 
 SEXP c_check_int(SEXP x, SEXP na_ok, SEXP lower, SEXP upper, SEXP tol, SEXP null_ok) {
     double dtol = asNumber(tol, "tol");
-    handle_na(x, na_ok);
-    handle_type_null(isIntegerish(x, dtol), "single integerish value", null_ok);
+    HANDLE_NA(x, na_ok);
+    HANDLE_TYPE_NULL(isIntegerish(x, dtol, FALSE), "single integerish value", null_ok);
     if (xlength(x) != 1)
         return result("Must have length 1");
-    assert(check_bounds(x, lower, upper));
+    ASSERT_TRUE(check_bounds(x, lower, upper));
     return ScalarLogical(TRUE);
 }
 
 SEXP c_check_number(SEXP x, SEXP na_ok, SEXP lower, SEXP upper, SEXP finite, SEXP null_ok) {
-    handle_na(x, na_ok);
-    handle_type_null(isStrictlyNumeric(x), "number", null_ok);
+    HANDLE_NA(x, na_ok);
+    HANDLE_TYPE_NULL(isStrictlyNumeric(x), "number", null_ok);
     if (xlength(x) != 1)
         return result("Must have length 1");
-    assert(check_vector_finite(x, finite));
-    assert(check_bounds(x, lower, upper));
+    ASSERT_TRUE(check_vector_finite(x, finite));
+    ASSERT_TRUE(check_bounds(x, lower, upper));
     return ScalarLogical(TRUE);
 }
 
 SEXP c_check_string(SEXP x, SEXP na_ok, SEXP min_chars, SEXP null_ok) {
-    handle_na(x, na_ok);
-    handle_type_null(isString(x), "string", null_ok);
+    HANDLE_NA(x, na_ok);
+    HANDLE_TYPE_NULL(isString(x), "string", null_ok);
     if (xlength(x) != 1)
         return result("Must have length 1");
     if (!isNull(min_chars)) {
@@ -531,9 +531,14 @@ SEXP c_check_string(SEXP x, SEXP na_ok, SEXP min_chars, SEXP null_ok) {
 }
 
 SEXP c_check_scalar(SEXP x, SEXP na_ok, SEXP null_ok) {
-    handle_na(x, na_ok);
-    handle_type_null(isVectorAtomic(x), "atomic scalar", null_ok);
+    HANDLE_NA(x, na_ok);
+    HANDLE_TYPE_NULL(isVectorAtomic(x), "atomic scalar", null_ok);
     if (xlength(x) != 1)
         return result("Must have length 1");
     return ScalarLogical(TRUE);
 }
+
+#undef HANDLE_TYPE
+#undef HANDLE_TYPE_NULL
+#undef HANDLE_NA
+#undef ASSERT_TRUE